Move clean string to its onw module
